requirements.txt
# Core Dependencies
requests>=2.31.0,<3.0.0
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=8.2.0,<9.0.0

# Development Dependencies
pytest>=7.4.0,<8.0.0
pytest-mock>=3.12.0,<4.0.0
responses>=0.24.0,<1.0.0
